Ignored comments during parsing
It was a nice idea to have comments, so that the comments could be
pushed out to the generated Mindustry Logic, but unfortunately, the
comments were causing more problems than they were worth. Removing
comments caused a bit more branching in the logic generation
(`null`-guards), but otherwise, the comments are not a pain anymore.

Fixes #22