Add getBlock to MINDUSTRY-TIPS-N-TRICKS.markdown
There are currently no code examples for the "getBlock" command. Many people, like me, at first tried to get the result of the command using "building = getBlock(x, y)", not seeing the result, they thought that Mindcode could not use such a command. Therefore, I consider it necessary to have an example of using this command.